More work on review comments
