+++ CMakeLists.txt
@@ -795,6 +795,8 @@ set(INPUT src/input/controller.cpp
           src/input/controller.h
           src/input/input_handler.cpp
           src/input/input_handler.h
+          src/input/input_mouse.cpp
+          src/input/input_mouse.h
 )
 
 set(EMULATOR src/emulator.cpp
